Bug 1695016 - Go back to 32 bits rendder task ids. r=gw
This is to work around fuzzers repeatedly finding test cases with 2^16 clip masks or other type of render tasks. Not something we would see in real pages nor something that we would handle well at all even with 32 bits ids but at least it won't get in the way of fuzzing.

Differential Revision: https://phabricator.services.mozilla.com/D116214

+++ gfx/wr/webrender/src/render_task_graph.rs
@@ -51,7 +51,7 @@ impl<'l> RenderTaskAllocation<'l> {
     #[inline(always)]
     pub fn init(self, value: RenderTask) -> RenderTaskId {
         RenderTaskId {
-            index: self.alloc.init(value) as u16,
+            index: self.alloc.init(value) as u32,
         }
     }
 }
@@ -61,12 +61,12 @@ impl<'l> RenderTaskAllocation<'l> {
 #[cfg_attr(feature = "capture", derive(Serialize))]
 #[cfg_attr(feature = "replay", derive(Deserialize))]
 pub struct RenderTaskId {
-    pub index: u16,
+    pub index: u32,
 }
 
 impl RenderTaskId {
     pub const INVALID: RenderTaskId = RenderTaskId {
-        index: u16::MAX,
+        index: u32::MAX,
     };
 }
 
@@ -264,7 +264,7 @@ impl RenderTaskGraphBuilder {
     pub fn add(&mut self) -> RenderTaskAllocation {
         // Assume every task is a root to start with
         self.roots.insert(
-            RenderTaskId { index: self.tasks.len() as u16 }
+            RenderTaskId { index: self.tasks.len() as u32 }
         );
 
         RenderTaskAllocation {
@@ -385,7 +385,7 @@ impl RenderTaskGraphBuilder {
         // Determine which pass each task can be freed on, which depends on which is
         // the last task that has this as an input.
         for i in 0 .. graph.tasks.len() {
-            let task_id = RenderTaskId { index: i as u16 };
+            let task_id = RenderTaskId { index: i as u32 };
             assign_free_pass(
                 task_id,
                 &mut self.child_task_buffer,
@@ -405,7 +405,7 @@ impl RenderTaskGraphBuilder {
         // Assign tasks to each pass based on their `render_on` attribute
         for (index, task) in graph.tasks.iter().enumerate() {
             if task.kind.is_a_rendering_operation() {
-                let id = RenderTaskId { index: index as u16 };
+                let id = RenderTaskId { index: index as u32 };
                 graph.passes[task.render_on.0].task_ids.push(id);
             }
         }
